Added the newest version of my PaleoMachineFramework
Already edited references.
Has support for the easy TileData system now! (Already implemented for
the Mixing Table)
Some changes in the framework itself (Should be more performant now and
solved a bug where the ItemStack in the input slot would have to be the
exact same size as the recipe one)
Improved Mixing Table.